Reworking of files to reduce redundancy and provide new additions
